requirements.txt
openpyxl==3.1.5
pandas>=1.5.0
Flask==3.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
pytest==7.4.0
pyTelegramBotAPI==4.14.0
requests==2.31.0
gunicorn==21.2.0
